+++ docs/changelog.rst
@@ -20,6 +20,7 @@ Changelog 📝
 **Improvements:**
 
 - `#417 <https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/417>`_: ``core``: move from ``semver`` to ``calver``.
+- `#423 <https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/423>`_: ``compute vm set custom-spec``: allow multiple ``--dns-suffix`` options for dns search settings.
 
 **Bug Fixes:**